requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
lyricsgenius>=3.0.0
pyarrow>=12.0.0
tqdm>=4.65.0 